# Conflicts:
#	.github/workflows/CI.yaml
#	dataland-backend/backendOpenApi.json
#	dataland-frontend/src/components/resources/dataTable/conversion/YesNoValueGetterFactory.ts
#	dataland-frontend/src/utils/Formatter.ts
#	dataland-frontend/src/utils/StringHumanizer.ts
#	settings.gradle.kts

+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/model/sfdr/categories/social/subcategories/SfdrSocialHumanRights.kt
@@ -3,7 +3,6 @@ package org.dataland.datalandbackend.model.sfdr.categories.social.subcategories
 import org.dataland.datalandbackend.model.datapoints.BaseDataPoint
 import org.dataland.datalandbackend.model.datapoints.ExtendedDataPoint
 import org.dataland.datalandbackend.model.enums.commons.YesNo
-import java.math.BigDecimal
 
 /**
  * --- API model ---
@@ -20,5 +19,5 @@ data class SfdrSocialHumanRights(
 
     val reportedForcedOrCompulsoryLabourIncidents: ExtendedDataPoint<YesNo>? = null,
 
-    val numberOfReportedIncidentsOfHumanRightsViolations: ExtendedDataPoint<BigDecimal>? = null,
+    val numberOfReportedIncidentsOfHumanRightsViolations: ExtendedDataPoint<Long>? = null,
 )
